--- pages/index.js
@@ -1,25 +1,0 @@
-import Head from 'next/head'
-// import Header from '@components/Header'
-// import Footer from '@components/Footer'
-import Link from 'next/link'
-
-export default function Home() {
-    return (
-        <div className="container">
-        {/* <Head>
-            <title>Next.js Starter!</title>
-            <link rel="icon" href="/favicon.ico" />
-        </Head>
-
-        <main>
-            <Header title="Welcome to my app!" />
-            <p className="description">
-            Get started by editing <code>pages/index.js</code>
-            </p>
-        </main>
-        <Link href="/page">Test page</Link>
-        <Link href="/logs">Logs Page</Link>
-        <Footer /> */}
-        </div>
-    )
-}
