--- prisma-unit-tests/client.ts
@@ -1,5 +1,0 @@
-// Prisma unit testing
-import { PrismaClient } from '@prisma/client'
-
-const prisma = new PrismaClient()
-export default prisma
