--- prisma-unit-tests/context.ts
@@ -1,17 +1,0 @@
-// Prisma unit testing
-import { PrismaClient } from '@prisma/client'
-import { mockDeep, DeepMockProxy } from 'jest-mock-extended'
-
-export type Context = {
-    prisma: PrismaClient
-}
-
-export type MockContext = {
-    prisma: DeepMockProxy<PrismaClient>
-}
-
-export const createMockContext = (): MockContext => {
-    return {
-        prisma: mockDeep<PrismaClient>(),
-    }
-}
