--- prisma-unit-tests/jest.config.js
@@ -1,7 +1,0 @@
-// Prisma unit testing
-module.exports = {
-    clearMocks: true,
-    preset: 'ts-jest',
-    testEnvironment: 'node',
-    setupFilesAfterEnv: ['<rootDir>/singleton.ts'],
-}
