--- prisma-unit-tests/singleton.ts
@@ -1,18 +1,0 @@
-// Prisma unit testing
-import { PrismaClient } from '@prisma/client'
-import { mockDeep, mockReset, DeepMockProxy } from 'jest-mock-extended'
-import { MockContext, Context, createMockContext } from './context'
-
-import prisma from './client'
-
-jest.mock('./client', () => ({
-    __esModule: true,
-    default: mockDeep<PrismaClient>(),
-}))
-
-beforeEach(() => {
-    mockReset(prismaMock)
-})
-
-
-export const prismaMock = prisma as unknown as DeepMockProxy<PrismaClient>
