--- styles/createLog.module.css
@@ -1,193 +1,0 @@
-.root {
-    --modal-size-xs: 320px;
-    --modal-size-sm: 380px;
-    --modal-size-md: 440px;
-    --modal-size-lg: 620px;
-    --modal-size-xl: 780px;
-    --modal-size: var(--modal-size-md);
-
-    --modal-y-offset: 5dvh;
-    --modal-x-offset: 5vw;
-
-    &[data-full-screen] {
-        --modal-border-radius: 0 !important;
-
-        .content {
-        --modal-content-flex: 0 0 100%;
-        --modal-content-max-height: auto;
-        --modal-content-height: 100dvh;
-        }
-
-        .inner {
-        --modal-inner-y-offset: 0;
-        --modal-inner-x-offset: 0;
-        }
-    }
-
-    &[data-centered] {
-        .inner {
-        --modal-inner-align: center;
-        }
-    }
-}
-
-.header {
-    border-start-start-radius: var(--modal-radius, var(--mantine-radius-default));
-    border-start-end-radius: var(--modal-radius, var(--mantine-radius-default));
-}
-
-.content {
-    flex: var(--modal-content-flex, 0 0 var(--modal-size));
-    max-width: 100%;
-    max-height: var(--modal-content-max-height, calc(100dvh - var(--modal-y-offset) * 2));
-    height: var(--modal-content-height, auto);
-    overflow-y: visible;
-
-    &[data-full-screen] {
-        border-radius: 0;
-    }
-
-    &[data-hidden] {
-        opacity: 0 !important;
-        pointer-events: none;
-    }
-}
-
-.form {
-    display: flex;
-    flex-direction: column;
-    gap: calc(var(--modal-y-offset) / 2);
-}
-.inner {
-    display: flex;
-    justify-content: center;
-    align-items: var(--modal-inner-align, flex-start);
-    padding-top: var(--modal-inner-y-offset, var(--modal-y-offset));
-    padding-bottom: var(--modal-inner-y-offset, var(--modal-y-offset));
-    padding-inline: var(--modal-inner-x-offset, var(--modal-x-offset));
-}
-
-.input {
-    width: 100%;
-    padding: 12px 20px;
-    border: none;
-    outline: none;
-}
-
-.inputDescription {
-    height: 150px;
-    display: inline-block;
-    width: var(--modal-size);
-    outline: 0;
-}
-
-.inputDescription:empty::before {
-    content: "Oh no what's happening?";
-    opacity: 0.6;
-}
-
-/* https://uiverse.io/zjssun/curly-seahorse-76 */
-.button {
-    width: fit-content;
-    position: relative; /*makes the sparkles not spaz out on page load*/
-    padding: 10px 22px;
-    border-radius: 6px;
-    border: none;
-    color: #fff;
-    cursor: pointer;
-    background-color: #7d2ae8;
-    transition: all 0.2s ease;
-}
-
-.button:active:enabled {
-  transform: scale(0.96);
-}
-
-.button:enabled:before,
-.button:enabled:after {
-    position: absolute;
-    content: "";
-    width: 150%;
-    left: 50%;
-    height: 100%;
-    transform: translateX(-50%);
-    z-index: -1000;
-    background-repeat: no-repeat;
-}
-
-.button:hover:enabled:before {
-    top: -70%;
-    background-image: radial-gradient(circle, #7d2ae8 20%, transparent 20%),
-        radial-gradient(circle, transparent 20%, #7d2ae8 20%, transparent 30%),
-        radial-gradient(circle, #7d2ae8 20%, transparent 20%),
-        radial-gradient(circle, #7d2ae8 20%, transparent 20%),
-        radial-gradient(circle, transparent 10%, #7d2ae8 15%, transparent 20%),
-        radial-gradient(circle, #7d2ae8 20%, transparent 20%),
-        radial-gradient(circle, #7d2ae8 20%, transparent 20%),
-        radial-gradient(circle, #7d2ae8 20%, transparent 20%),
-        radial-gradient(circle, #7d2ae8 20%, transparent 20%);
-    background-size: 10% 10%, 20% 20%, 15% 15%, 20% 20%, 18% 18%, 10% 10%, 15% 15%,
-        10% 10%, 18% 18%;
-    background-position: 50% 120%;
-    animation: greentopBubbles 0.6s ease;
-}
-
-.button:disabled,
-.button[disabled]{
-    opacity: 50%;
-}
-
-
-/* .reactSelect [class$="-control"]{
-    background-color: var(--background-1);
-} */
-
-@keyframes greentopBubbles {
-    0% {
-        background-position: 5% 90%, 10% 90%, 10% 90%, 15% 90%, 25% 90%, 25% 90%,
-        40% 90%, 55% 90%, 70% 90%;
-    }
-
-    50% {
-        background-position: 0% 80%, 0% 20%, 10% 40%, 20% 0%, 30% 30%, 22% 50%,
-        50% 50%, 65% 20%, 90% 30%;
-    }
-
-    100% {
-        background-position: 0% 70%, 0% 10%, 10% 30%, 20% -10%, 30% 20%, 22% 40%,
-        50% 40%, 65% 10%, 90% 20%;
-        background-size: 0% 0%, 0% 0%, 0% 0%, 0% 0%, 0% 0%, 0% 0%;
-    }
-}
-
-.button:hover::after {
-    bottom: -70%;
-    background-image: radial-gradient(circle, #7d2ae8 20%, transparent 20%),
-        radial-gradient(circle, #7d2ae8 20%, transparent 20%),
-        radial-gradient(circle, transparent 10%, #7d2ae8 15%, transparent 20%),
-        radial-gradient(circle, #7d2ae8 20%, transparent 20%),
-        radial-gradient(circle, #7d2ae8 20%, transparent 20%),
-        radial-gradient(circle, #7d2ae8 20%, transparent 20%),
-        radial-gradient(circle, #7d2ae8 20%, transparent 20%);
-    background-size: 15% 15%, 20% 20%, 18% 18%, 20% 20%, 15% 15%, 20% 20%, 18% 18%;
-    background-position: 50% 0%;
-    animation: greenbottomBubbles 0.6s ease;
-}
-
-@keyframes greenbottomBubbles {
-    0% {
-        background-position: 10% -10%, 30% 10%, 55% -10%, 70% -10%, 85% -10%,
-        70% -10%, 70% 0%;
-    }
-
-    50% {
-        background-position: 0% 80%, 20% 80%, 45% 60%, 60% 100%, 75% 70%, 95% 60%,
-        105% 0%;
-    }
-
-    100% {
-        background-position: 0% 90%, 20% 90%, 45% 70%, 60% 110%, 75% 80%, 95% 70%,
-        110% 10%;
-        background-size: 0% 0%, 0% 0%, 0% 0%, 0% 0%, 0% 0%, 0% 0%;
-    }
-}
