--- styles/logs.module.css
@@ -1,78 +1,0 @@
-
-  
-  /* .body {
-    padding-left: 54px;
-    padding-top: var(--mantine-spacing-sm);
-    font-size: var(--mantine-font-size-sm);
-  } */
-
-   
-.body{
-    padding: 1.5rem 0;
-    margin: auto;
-    display: flex;
-    flex-direction: column;
-    max-width: 700px;
-    gap: 1.5rem;
-    /* background-color: var(--background-1); */
-}
-
-.content {
-    & > p:last-child {
-        margin-bottom: 0;
-    }
-}
-
-.comment {
-    padding: var(--mantine-spacing-lg) var(--mantine-spacing-xl);
-    display: flex;
-    flex-wrap: wrap;
-    /* background-color: var(--background-1); */
-    background-color: #f8fafc;
-    /* border-color: var(--background-2); */
-    .text {
-        font-size: small;
-        /* color: var(--text); */
-    }
-}
-
-.loader{
-    /* display: block;
-    margin: auto; */
-    /* position: absolute; 
-    top: 50%; 
-    left: 50%; */
-    align-self: center;
-    margin-top: 3rem;
-}
-
-
-.menu {
-    margin-left: auto;
-    /* color: var(--text); */
-    color: black;
-    opacity: 0.6;
-}
-
-.dropdown {
-    /* background: var(--background-1); */
-    background: #f8fafc;
-}
-
-.tags {
-    /* className="space-y-2" style={{listStyleType:'none', display:'flex'}} */
-    list-style-type: none;
-    display: flex;
-    align-self: flex-end;
-    flex: 0 0 100%; /* flex-grow, flex-shrink, flex-basis */
-    justify-content: flex-end;
-}
-
-.tag {
-    margin: 0 3px;
-}
-
-.bar {
-    display: flex;
-    padding: 1rem;
-}
